Envelope

+++ server/s_source.hpp
@@ -7,7 +7,7 @@ class Source : public Entity
 public:
   Source(class Sink &);
   virtual ~Source();
-  virtual auto pull(int samples) -> std::vector<int16_t> = 0;
+  virtual auto pull(int samples) -> std::vector<float> = 0;
 
 protected:
   std::reference_wrapper<Sink> sink;
